requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.3.0

# Storage
boto3==1.34.0
aiofiles==23.2.1

# AI/ML
openai==1.6.1
anthropic==0.8.1

# Task Queue
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# HTTP Client
httpx==0.25.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
python-dotenv==1.0.0
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# Monitoring
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7